pyproject.toml
[tool.poetry]
name = "flask-project"
version = "0.1.0"
description = ""
authors = ["tolik <9779930@mail.ru>"]
readme = "README.md"
packages = [{include = "flask_project"}]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^2.2.3"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
psycopg2-binary = "^2.9.6"
wtforms = "^3.0.1"
email-validator = "^2.0.0.post2"
flask-wtf = "^1.1.1"
flask-admin = "^1.6.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
